WIP: Start refactoring validationAPI
Use a shared context object (self.ctx)
